test: increase coverage
