+++ packages/server/package.json
@@ -9,8 +9,8 @@
         "postinstall": "link-module-alias",
         "production": "NODE_ENV=production node src/main.js",
         "start": "NODE_ENV=development nodemon src/main.js",
-        "test": "NODE_ENV=test npm run reset-and-seed-db && lab -a code -t 80 -v -I 'core,__core-js_shared__'",
-        "test-coverage": "NODE_ENV=test lab -c -a code -r html -o coverage/coverage.html",
+        "test": "NODE_ENV=test npm run reset-and-seed-db && lab -a code -t 80 -v -I 'core,__core-js_shared__' test/index.js",
+        "test-coverage": "NODE_ENV=test lab -c -a code -r html -o coverage/coverage.html test/index.js",
         "reset-db": "npx knex migrate:rollback && npx knex migrate:latest",
         "seed-db": "npx knex seed:run",
         "reset-and-seed-db": "npx knex migrate:rollback && npx knex migrate:latest && npx knex seed:run",
