+++ packages/server/src/data/createAppAsync.js
@@ -1,5 +1,16 @@
 const uuid = require('uuid/v4')
 
+const joi = require('joi')
+
+const { AppType } = require('@enum')
+
+const paramsSchema = joi.object().keys({
+    userId: joi.number().required(),
+    orgId: joi.number().required(),
+    appType: joi.string().required.valid(AppType),
+}).options({ allowUnknown: true })
+
+
 /**
  * Creates an app and returns the database id
  * @param {number} userId
@@ -10,26 +21,14 @@ const uuid = require('uuid/v4')
  */
 const createAppAsync = async (params, knex, transaction) =>  {
 
+    const validation = joi.validate(params, paramsSchema)
+
+    if ( validation.error !== null ) {
+        throw new Error(validation.error)
+    }
+
     const { userId, orgId, appType } = params
-    if ( !userId ) {
-        throw new Error(`Parameter missing: 'userId'`)
-    }
 
-    if ( !orgId ) {
-        throw new Error(`Parameter missing: 'orgId'`)
-    }
-
-    if ( !appType ) {
-        throw new Error(`Parameter missing: 'appType''`)
-    }
-
-    if ( !knex ) {
-        throw new Error(`Parameter missing: 'knex'`)
-    }
-
-    if ( !transaction ) {
-        throw new Error(`Parameter missing: 'transaction'`)
-    }
 
     //generate a new uuid to insert
     const appUuid = uuid()
