+++ packages/server/src/security/index.js
@@ -1,6 +1,6 @@
 /**
  * This returns true if the request is authenticated (e.g. contains a valid token)
- * @param {*} request 
+ * @param {*} request
  */
 const isAuthenticated = (request) => {
 
@@ -13,8 +13,8 @@ const isAuthenticated = (request) => {
 
 /**
  * Returns true if the JWT/user for the request contains a role with the specified name
- * @param {*} request 
- * @param {string} role 
+ * @param {*} request
+ * @param {string} role
  */
 const hasRole = (request, role) => {
 
@@ -27,36 +27,36 @@ const hasRole = (request, role) => {
 
 /**
  * Checks if the user on the request has permissions to delete an app
- * @param {*} request 
- * @param {*} hapi 
+ * @param {*} request
+ * @param {*} hapi
  */
 const canDeleteApp = (request, hapi) => isAuthenticated(request) && hasRole(request, 'ROLE_MANAGER')
 
 /**
  * Checks if the user on the request has permissions to change the status of an app
- * @param {*} request 
- * @param {*} hapi 
+ * @param {*} request
+ * @param {*} hapi
  */
 const canChangeAppStatus = (request, hapi) => isAuthenticated(request) && hasRole(request, 'ROLE_MANAGER')
 
 /**
  * Checks if the user on the request has permissions to create an app version
- * @param {*} request 
- * @param {*} hapi 
+ * @param {*} request
+ * @param {*} hapi
  */
 const canCreateAppVersion = (request, hapi) => isAuthenticated(request)
 
 /**
  * Checks if the user on the request has permissions to create an app
- * @param {*} request 
- * @param {*} hapi 
+ * @param {*} request
+ * @param {*} hapi
  */
 const canCreateApp = (request, hapi) => isAuthenticated(request)
 
 /**
  * Checks if the user on the request has permissions to see all apps
- * @param {*} request 
- * @param {*} hapi 
+ * @param {*} request
+ * @param {*} hapi
  */
 const canSeeAllApps = (request, hapi) => isAuthenticated(request) && hasRole(request, 'ROLE_MANAGER')
